% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialsimulation-package.R
\name{SIM}
\alias{SIM}
\title{Performs a simple trial simulation}
\usage{
SIM(...)
}
\arguments{
\item{nMin}{minimum to trigger the analysis}

\item{nMax}{upper limit of the sample size}

\item{nByVec}{acrual rate}

\item{phi}{Substantial Evidence Threshold}

\item{targetDelta}{Hypothesized treatment effect}

\item{block0}{Randomization information}

\item{enr_mn}{mean, one patient per 4 weeks}

\item{enr_sd}{standard deviation}

\item{asc_mn}{mean, 8 weeks +/- 1 week}

\item{asc_sd}{standard deviation}

\item{skp_mn}{mean of skeptical prior}

\item{true_sd}{true standard deviation}

\item{true_mn}{true mean change}

\item{hst_n}{sample size}

\item{hst_mn}{mean (also mean of normal power prior)}

\item{hst_sd}{standard deviation}

\item{nSims}{number of simulations}

\item{nSamp}{number of MCMC samples}
}
\value{
a data frame with results
\itemize{
\item \emph{skp_tau}   -- inverse of standard deviation
\item \emph{skp_mu1}   -- effect of control arm
\item \emph{skp_mu0}   -- effect of treated arm
\item \emph{skp_diff}  -- treatment effect
\item \emph{skp_pp}    -- posterior probability of activity
\item \emph{ent_tau}   -- inverse of standard deviation
\item \emph{ent_mu1}   -- effect of control arm
\item \emph{ent_mu0}   -- effect of treated arm
\item \emph{ent_diff}  -- treatment effect
\item \emph{ent_pp}    -- posterior probability of activity
\item \emph{analysis}  -- number of analysis performed
\item \emph{nInt}      -- Number of patients at interim where early stoppage takes place
\item \emph{nOngoing}  -- Number of patients at interim where early stoppage doesn't take place
\item \emph{nFin}      -- Number of patients currently enrolled
\item \emph{essFin}    -- effective sample size
\item \emph{y1Fin}     -- estimated effect of control group
\item \emph{y2Fin}     -- estimated treatment effect
\item \emph{fut}       -- percentage of futility stopping (interim only)
\item \emph{effInt}    -- percentage of efficacy stopping (interim)
\item \emph{effFin}    -- percentage of efficacy stopping (final)
\item \emph{true_sd}   -- true standard deviation in data generation model
\item \emph{true_mu1}  -- true mean change in control arm
\item \emph{true_mu0}  -- true mean change in treated arm
\item \emph{a0}        -- deterministic power prior parameter
}
}
\description{
Performs a simple trial simulation
}
\examples{
## Set simulation parameters
hst_n  = 25.0
hst_mn =  0.0
hst_sd =  8.0
nSims =   500
nSamp =   500
nMin   =   24
nMax   =   70
nByVec = c(nMin,8,8,8,8,14)
phi    = 0.025
targetDelta = 5.5
block  = c(1,1,2,2)   ## note 1 = treatment, 2 = control
block_size = length(block)
enr_mn = 4.0
enr_sd = 0.5
asc_mn = 8.00  ## 8 weeks +/- 1 week
asc_sd = 0.25
skp_mn = 0
ent_mn = targetDelta
true_sd = 8.0
tmu1 = 0.0
tmu2 = 5.5
true_mn = c(tmu1,tmu2)

set.seed(1)
sims = SIM(nMin, nMax, nByVec, phi, targetDelta, block, enr_mn, enr_sd, asc_mn, asc_sd, skp_mn, 
           true_sd, true_mn, hst_n, hst_mn, hst_sd, nSims, nSamp)

}
